// src/app/extensions/quoting/models/quote-request-item/quote-request-item.interface.ts
import { Price } from 'ish-core/models/price/price.model';

export interface QuoteRequestItemData {
  type: string;
  quantity: {
    type: string;
    value: number;
    unit?: string;
  };
  singlePrice: Price;
  totalPrice: Price;
  originQuantity: {
    type: string;
    value: number;
    unit?: string;
  };
  originSinglePrice: Price;
  originTotalPrice: Price;
  productSKU: string;
}